decouple generation of build plan from docker image creation
