+++ src/lib.rs
@@ -1,20 +1,26 @@
 use std::env;
 
-use crate::{
-    nixpacks::{
-        app::App, environment::Environment, logger::Logger, nix::Pkg, plan::BuildPlan, AppBuilder,
-        AppBuilderOptions,
+use anyhow::{bail, Result};
+use nixpacks::{
+    builder::{
+        docker::{DockerBuilder, DockerBuilderOptions},
+        Builder,
     },
+    plan::{
+        generator::{GeneratePlanOptions, NixpacksBuildPlanGenerator},
+        BuildPlan, PlanGenerator,
+    },
+};
+use {
+    nixpacks::{app::App, environment::Environment, logger::Logger, nix::pkg::Pkg},
     providers::{
         crystal::CrystalProvider, deno::DenoProvider, go::GolangProvider,
         haskell::HaskellStackProvider, node::NodeProvider, python::PythonProvider,
-        rust::RustProvider,
+        rust::RustProvider, Provider,
     },
 };
-use anyhow::{bail, Result};
-use providers::Provider;
 
-pub(crate) mod chain;
+mod chain;
 pub mod nixpacks;
 pub mod providers;
 
@@ -30,119 +36,105 @@ pub fn get_providers() -> Vec<&'static dyn Provider> {
     ]
 }
 
-pub fn gen_plan(
+pub fn generate_build_plan(
     path: &str,
-    custom_pkgs: Vec<&str>,
-    custom_build_cmd: Option<String>,
-    custom_start_cmd: Option<String>,
     envs: Vec<&str>,
-    pin_pkgs: bool,
+    plan_options: &GeneratePlanOptions,
 ) -> Result<BuildPlan> {
-    let logger = Logger::new();
-    let providers = get_providers();
-
-    let options = AppBuilderOptions {
-        custom_pkgs: custom_pkgs.iter().map(|p| Pkg::new(p)).collect(),
-        custom_build_cmd,
-        custom_start_cmd,
-        pin_pkgs,
-        out_dir: None,
-        plan_path: None,
-        tags: Vec::new(),
-        labels: Vec::new(),
-        quiet: false,
-    };
-
     let app = App::new(path)?;
-    let environment = create_environment(envs)?;
-    let mut app_builder = AppBuilder::new(None, &app, &environment, &logger, &options)?;
+    let environment = Environment::from_envs(envs)?;
 
-    let plan = app_builder.plan(providers)?;
+    let mut generator = NixpacksBuildPlanGenerator::new(get_providers(), plan_options.to_owned());
+    let plan = generator.generate_plan(&app, &environment)?;
+
     Ok(plan)
 }
 
-#[allow(clippy::too_many_arguments)]
-pub fn build(
+pub fn create_docker_image(
     path: &str,
-    name: Option<String>,
-    custom_pkgs: Vec<&str>,
-    custom_build_cmd: Option<String>,
-    custom_start_cmd: Option<String>,
-    pin_pkgs: bool,
     envs: Vec<&str>,
-    plan_path: Option<String>,
-    out_dir: Option<String>,
-    tags: Vec<&str>,
-    labels: Vec<&str>,
-    quiet: bool,
+    plan_options: &GeneratePlanOptions,
+    build_options: &DockerBuilderOptions,
 ) -> Result<()> {
-    let logger = Logger::new();
-    let providers = get_providers();
-
-    let options = AppBuilderOptions {
-        custom_pkgs: custom_pkgs.iter().map(|p| Pkg::new(p)).collect(),
-        custom_build_cmd,
-        custom_start_cmd,
-        pin_pkgs,
-        out_dir,
-        plan_path,
-        tags: tags.iter().map(|s| s.to_string()).collect(),
-        labels: labels.iter().map(|s| s.to_string()).collect(),
-        quiet,
-    };
-
     let app = App::new(path)?;
-    let environment = create_environment(envs)?;
-    let mut app_builder = AppBuilder::new(name, &app, &environment, &logger, &options)?;
+    let environment = Environment::from_envs(envs)?;
 
-    app_builder.build(providers)?;
+    let mut generator = NixpacksBuildPlanGenerator::new(get_providers(), plan_options.to_owned());
+    let plan = generator.generate_plan(&app, &environment)?;
+
+    let logger = Logger::new();
+    let builder = DockerBuilder::new(logger, build_options.to_owned());
+    builder.create_image(&app, &plan)?;
 
     Ok(())
 }
 
-pub fn create_environment(envs: Vec<&str>) -> Result<Environment> {
-    let mut environment = Environment::default();
-    for env in envs {
-        let v: Vec<&str> = env.split('=').collect();
-        if v.len() == 1 {
-            // Pull the variable from the current environment
-            let name = v[0];
-            if let Ok(value) = env::var(name) {
-                // Variable is set
-                environment.set_variable(name.to_string(), value);
-            }
-        } else if v.len() > 2 {
-            bail!("Unable to parse variable string");
-        } else {
-            // Use provided name, value pair
-            environment.set_variable(v[0].to_string(), v[1].to_string());
-        }
-    }
+// pub fn gen_plan(
+//     path: &str,
+//     custom_pkgs: Vec<&str>,
+//     custom_build_cmd: Option<String>,
+//     custom_start_cmd: Option<String>,
+//     envs: Vec<&str>,
+//     pin_pkgs: bool,
+// ) -> Result<BuildPlan> {
+//     let logger = Logger::new();
+//     let providers = get_providers();
 
-    Ok(environment)
-}
+//     let options = AppBuilderOptions {
+//         custom_pkgs: custom_pkgs.iter().map(|p| Pkg::new(p)).collect(),
+//         custom_build_cmd,
+//         custom_start_cmd,
+//         pin_pkgs,
+//         out_dir: None,
+//         plan_path: None,
+//         tags: Vec::new(),
+//         labels: Vec::new(),
+//         quiet: false,
+//     };
 
-#[cfg(test)]
-mod tests {
-    use crate::create_environment;
+//     let app = App::new(path)?;
+//     let environment = create_environment(envs)?;
+//     let mut app_builder = AppBuilder::new(None, &app, &environment, &logger, &options)?;
 
-    #[test]
-    fn test_environment_variable_parsing() {
-        let environment =
-            create_environment(vec!["HELLO=world", "CARGO_PKG_NAME", "NON_EXISTANT"]).unwrap();
-        assert_eq!(
-            environment.get_variable("HELLO"),
-            Some(&"world".to_string())
-        );
-        assert_eq!(
-            environment.get_variable("CARGO_PKG_NAME"),
-            Some(&"nixpacks".to_string())
-        );
-        assert!(environment.get_variable("NON_EXISTANT").is_none());
-    }
+//     let plan = app_builder.plan(providers)?;
+//     Ok(plan)
+// }
 
-    #[test]
-    fn test_create_invalid_environment() {
-        assert!(create_environment(vec!["INVALID=ENV=CONFIG"]).is_err());
-    }
-}
+// #[allow(clippy::too_many_arguments)]
+// pub fn build(
+//     path: &str,
+//     name: Option<String>,
+//     custom_pkgs: Vec<&str>,
+//     custom_build_cmd: Option<String>,
+//     custom_start_cmd: Option<String>,
+//     pin_pkgs: bool,
+//     envs: Vec<&str>,
+//     plan_path: Option<String>,
+//     out_dir: Option<String>,
+//     tags: Vec<&str>,
+//     labels: Vec<&str>,
+//     quiet: bool,
+// ) -> Result<()> {
+//     let logger = Logger::new();
+//     let providers = get_providers();
+
+//     let options = AppBuilderOptions {
+//         custom_pkgs: custom_pkgs.iter().map(|p| Pkg::new(p)).collect(),
+//         custom_build_cmd,
+//         custom_start_cmd,
+//         pin_pkgs,
+//         out_dir,
+//         plan_path,
+//         tags: tags.iter().map(|s| s.to_string()).collect(),
+//         labels: labels.iter().map(|s| s.to_string()).collect(),
+//         quiet,
+//     };
+
+//     let app = App::new(path)?;
+//     let environment = create_environment(envs)?;
+//     let mut app_builder = AppBuilder::new(name, &app, &environment, &logger, &options)?;
+
+//     app_builder.build(providers)?;
+
+//     Ok(())
+// }
