+++ src/nixpacks/environment.rs
@@ -1,4 +1,5 @@
-use std::collections::HashMap;
+use anyhow::{bail, Result};
+use std::{collections::HashMap, env};
 
 pub type EnvironmentVariables = HashMap<String, String>;
 
@@ -10,6 +11,28 @@ pub struct Environment {
 impl Environment {
     pub fn new(variables: EnvironmentVariables) -> Environment {
         Environment { variables }
+    }
+
+    pub fn from_envs(envs: Vec<&str>) -> Result<Environment> {
+        let mut environment = Environment::default();
+        for env in envs {
+            let v: Vec<&str> = env.split('=').collect();
+            if v.len() == 1 {
+                // Pull the variable from the current environment
+                let name = v[0];
+                if let Ok(value) = env::var(name) {
+                    // Variable is set
+                    environment.set_variable(name.to_string(), value);
+                }
+            } else if v.len() > 2 {
+                bail!("Unable to parse variable string");
+            } else {
+                // Use provided name, value pair
+                environment.set_variable(v[0].to_string(), v[1].to_string());
+            }
+        }
+
+        Ok(environment)
     }
 
     pub fn get_variable(&self, name: &str) -> Option<&String> {
@@ -54,5 +77,25 @@ mod tests {
             environment.get_variable("hello"),
             Some(&"world".to_string())
         );
+    }
+
+    #[test]
+    fn test_environment_variable_parsing() {
+        let environment =
+            Environment::from_envs(vec!["HELLO=world", "CARGO_PKG_NAME", "NON_EXISTANT"]).unwrap();
+        assert_eq!(
+            environment.get_variable("HELLO"),
+            Some(&"world".to_string())
+        );
+        assert_eq!(
+            environment.get_variable("CARGO_PKG_NAME"),
+            Some(&"nixpacks".to_string())
+        );
+        assert!(environment.get_variable("NON_EXISTANT").is_none());
+    }
+
+    #[test]
+    fn test_create_invalid_environment() {
+        assert!(Environment::from_envs(vec!["INVALID=ENV=CONFIG"]).is_err());
     }
 }
